debug, add deleteDirectory implementation(win32), add rfc 775 cmds
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2551 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/source/connection.cpp
@@ -41,7 +41,7 @@ Connection::Connection()
   connectionBuffer = new   char [MYSERVER_KB(8)];
   protocolBuffer = 0;
   socket = 0;
-  priority = -1;
+  priority = (u_long)-1;
   dataRead = 0;
 }
